Set up header layout
Use auto margin trick to push the page title to
the left and the links to the right.

Use flex-box to fit the two boxes in
.content_box_one evenly.

Add image to the right side of box one.

Edit styles of sign up button.

This replicates the header design of the project
while adding my own content.